+++ chat_test.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/GetStream/getstream-go/v4"
+	"github.com/GetStream/getstream-go/v5"
 	"github.com/stretchr/testify/require"
 )
 
@@ -112,6 +112,13 @@ func TestChatDeleteChannel(t *testing.T) {
 	require.NoError(t, err)
 
 	_, err = client.Chat().DeleteChannel(context.Background(), "", "", &getstream.DeleteChannelRequest{})
+	require.NoError(t, err)
+}
+func TestChatGetChannel(t *testing.T) {
+	client, err := getstream.NewClient("key", "secret", getstream.WithHTTPClient(&StubHTTPClient{}))
+	require.NoError(t, err)
+
+	_, err = client.Chat().GetChannel(context.Background(), "", "", &getstream.GetChannelRequest{})
 	require.NoError(t, err)
 }
 func TestChatUpdateChannelPartial(t *testing.T) {